refactor(authentication,chat,database,email,forms,push-notifications,sms,storage): admin files
