+++ modules/authentication/src/Authentication.ts
@@ -11,7 +11,7 @@ import path from 'path';
 import { isNil } from 'lodash';
 import { status } from '@grpc/grpc-js';
 import AppConfigSchema, { Config } from './config';
-import { AdminHandlers } from './admin/admin';
+import { AdminHandlers } from './admin';
 import { AuthenticationRoutes } from './routes/routes';
 import * as models from './models';
 import { ISignTokenOptions } from './interfaces/ISignTokenOptions';
