+++ modules/push-notifications/src/PushNotifications.ts
@@ -5,7 +5,7 @@ import {
   HealthCheckStatus,
 } from '@conduitplatform/grpc-sdk';
 import AppConfigSchema, { Config } from './config';
-import { AdminHandlers } from './admin/admin';
+import { AdminHandlers } from './admin';
 import { PushNotificationsRoutes } from './routes/routes';
 import * as models from './models';
 import {
